requirements.txt
# RAG System Dependencies
# LangChain Core
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# LangChain Integrations
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
langchain-pinecone>=0.1.0

# Vector Store
pinecone>=5.0.0

# Embeddings & LLM Providers
openai>=1.0.0
google-generativeai>=0.8.0

# Utilities
python-dotenv>=1.0.0
